requirements.txt
selenium==4.17.2
pytest==8.0.0
requests==2.31.0
mysql-connector-python==8.3.0